+++ packages/atomic-angular/projects/atomic-angular/src/lib/stencil-generated/components.ts
@@ -362,13 +362,13 @@ export declare interface AtomicIconButton extends Components.AtomicIconButton {}
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['clickCallback', 'icon', 'labelI18nKey', 'tooltip']
+  inputs: ['buttonRef', 'clickCallback', 'icon', 'labelI18nKey', 'tooltip']
 })
 @Component({
   selector: 'atomic-icon-button',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['clickCallback', 'icon', 'labelI18nKey', 'tooltip']
+  inputs: ['buttonRef', 'clickCallback', 'icon', 'labelI18nKey', 'tooltip']
 })
 export class AtomicIconButton {
   protected el: HTMLElement;
@@ -419,6 +419,47 @@ export class AtomicInsightHistoryToggle {
     this.el = r.nativeElement;
   }
 }
+
+
+export declare interface AtomicInsightRefineModal extends Components.AtomicInsightRefineModal {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['isOpen', 'openButton']
+})
+@Component({
+  selector: 'atomic-insight-refine-modal',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['isOpen', 'openButton']
+})
+export class AtomicInsightRefineModal {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
+export declare interface AtomicInsightRefineToggle extends Components.AtomicInsightRefineToggle {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined
+})
+@Component({
+  selector: 'atomic-insight-refine-toggle',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>'
+})
+export class AtomicInsightRefineToggle {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
 
 export declare interface AtomicInsightResult extends Components.AtomicInsightResult {}
 
@@ -610,7 +651,7 @@ export class AtomicNumericRange {
 
 export declare interface AtomicPager extends Components.AtomicPager {
   /**
-   *
+   *  
    */
   'atomic/scrollToTop': EventEmitter<CustomEvent<any>>;
 
@@ -1480,7 +1521,7 @@ export class AtomicSmartSnippet {
 
 export declare interface AtomicSmartSnippetFeedbackModal extends Components.AtomicSmartSnippetFeedbackModal {
   /**
-   *
+   *  
    */
   'feedbackSent': EventEmitter<CustomEvent<any>>;